+++ src/apis/security/client.ts
@@ -20,6 +20,30 @@ export interface UpdatePasswordForm {
   newPassword: string;
 }
 
+export type RefreshTokenRequest = RequestWithBody<RefreshTokenForm>;
+
+export interface RefreshTokenForm {
+  refreshToken: string;
+}
+
+export type InitializeForgotPasswordRequest = RequestWithBody<
+  InitializeForgotPasswordForm
+>;
+
+export interface InitializeForgotPasswordForm {
+  username: string;
+}
+
+export type ConfirmForgotPasswordRequest = RequestWithBody<
+  ConfirmForgotPasswordForm
+>;
+
+export interface ConfirmForgotPasswordForm {
+  username: string;
+  newPassword: string;
+  confirmationCode: string;
+}
+
 export class SecurityClient extends BaseClient {
   constructor(config: ClientConfig, axiosInstance: AxiosInstance = axios) {
     super(config, axiosInstance);
@@ -41,6 +65,37 @@ export class SecurityClient extends BaseClient {
     const url = `${config.host}/security/confirm-account`;
 
     return this.axiosInstance.post<LoginResponse>(url, request.body, {
+      headers: config.headers,
+    });
+  }
+
+  refresh(request: RefreshTokenRequest): Promise<AxiosResponse<LoginResponse>> {
+    const config = this.config.merge(request as PortfolioRequest);
+    const url = `${config.host}/security/refresh`;
+
+    return this.axiosInstance.post<LoginResponse>(url, request.body, {
+      headers: config.headers,
+    });
+  }
+
+  initForgotPassword(
+    request: InitializeForgotPasswordRequest
+  ): Promise<AxiosResponse<PortfolioResponse>> {
+    const config = this.config.merge(request as PortfolioRequest);
+    const url = `${config.host}/security/init-forgot-password`;
+
+    return this.axiosInstance.post<PortfolioResponse>(url, request.body, {
+      headers: config.headers,
+    });
+  }
+
+  confirmForgotPassword(
+    request: ConfirmForgotPasswordRequest
+  ): Promise<AxiosResponse<PortfolioResponse>> {
+    const config = this.config.merge(request as PortfolioRequest);
+    const url = `${config.host}/security/confirm-forgot-password`;
+
+    return this.axiosInstance.post<PortfolioResponse>(url, request.body, {
       headers: config.headers,
     });
   }